+++ pkg/controller/ads/ads_event.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package envoy
+package ads
 
 import (
 	"fmt"
