+++ pkg/controller/ads/ads_stream.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package envoy
+package ads
 
 import (
 	"context"
