+++ pkg/controller/ads/convert_filter_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package envoy
+package ads
 
 import (
 	"testing"
